requirements.txt
torch
numpy
scikit-learn
easydict
pyyaml
wandb
matplotlib
cython
POT
